feat : 당첨 번호 입력 & 당첨 번호 예외 처리

+++ src/main/kotlin/lotto/domain/Purchase.kt
@@ -14,7 +14,7 @@ class Purchase {
     }
 
     fun getLottoCountFromAmount(): Int {
-        val amount = Input.inputAmount().toInt()
+        val amount = Input.inputAmount()
         checkValidationAmount(amount)
         return amount / Constant.PRICE_PER_LOTTO
     }
